renamet fra searchfield -> search
