--- @navikt/core/react/src/form/search-field/index.ts
@@ -1,2 +1,0 @@
-export { default as SearchField, SearchFieldProps } from "./SearchField";
-export * from "./SearchField";
