+++ @navikt/core/react/src/form/search/index.ts
@@ -1,0 +1,1 @@
+export { default as Search, SearchProps, SearchClearEvent } from "./Search";
