+++ @navikt/core/react/src/form/search/useSearch.ts
